[CCUBE-252][KT] Create component to render file items. Preparation for part 3
